Add files via upload
A portfolio website created using HTML and CSS server as a digital showcase for an individual's skills and projects. It typically includes a structured layout with section such as Home,  About,  Resume, and Contact. The website is built with HTML for semantic structure and CSS for styling. Key elements include a navigation bar, a personal bio, education, a project displays, and a contact section to facilitate communication.